# python/server/swagger_server/models/action_result.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from server.swagger_server.models.base_model_ import Model
from server.swagger_server import util


class ActionResult(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, id: str=None, data: str=None):  # noqa: E501
        """ActionResult - a model defined in Swagger

        :param id: The id of this ActionResult.  # noqa: E501
        :type id: str
        :param data: The data of this ActionResult.  # noqa: E501
        :type data: str
        """
        self.swagger_types = {
            'id': str,
            'data': str
        }

        self.attribute_map = {
            'id': 'id',
            'data': 'data'
        }

        self._id = id
        self._data = data

    @classmethod
    def from_dict(cls, dikt) -> 'ActionResult':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The ActionResult of this ActionResult.  # noqa: E501
        :rtype: ActionResult
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> str:
        """Gets the id of this ActionResult.

        Docker container id  # noqa: E501

        :return: The id of this ActionResult.
        :rtype: str
        """
        return self._id

    @id.setter
    def id(self, id: str):
        """Sets the id of this ActionResult.

        Docker container id  # noqa: E501

        :param id: The id of this ActionResult.
        :type id: str
        """

        self._id = id

    @property
    def data(self) -> str:
        """Gets the data of this ActionResult.


        :return: The data of this ActionResult.
        :rtype: str
        """
        return self._data

    @data.setter
    def data(self, data: str):
        """Sets the data of this ActionResult.


        :param data: The data of this ActionResult.
        :type data: str
        """

        self._data = data